split off registry proxying into proxy module
The abstraction is not good, the code is buggy.
But it will do until it is re-done to use the SQL database.
